[css-text] Consider breaking opportunities of inline siblings
First of all, bear in mind that this change affects only to
intrinsic size computation in legacy layout.

During the intrinsic size computation of blocks with inline children,
we are determining the breaking opportunities for each child. When the
word-break: break-word is used, each of this breaking opportunities
should account for the min-content size.

We only consider breakable locations for start and end if they are
br elements or spaces, under auto-wrap, to compute the intrinsic size
of a box. This is fine, since we are already using a break iterator
to determine the min-size of each word. However, we were incorrectly
summing the first_line_min_width for each inline sibling, even when
they were part of the same text line.

This change tries to avoid that problem by setting the child's
min-width as the box's min_logical_width before adding the first line
min-width amount. That's what we are doing in case there is a breakable
location at the end of the line.

Bug: 1013644
Change-Id: I04261b323cd029a624724363a566fccc826863af