Pass outside fetch client settings object for ServiceWorkerContainer#register()
A brief summary of this plumbing is below:

blink::ServiceWorkerContainer::registerServiceWorker()
-> (IPC from a renderer to the browser)
-> content::ServiceworkerContextCore::RegisterServiceWorker()
   Start a registration job asynchronously
   Create a ServiceWorkerVersion and set the passed settings object
-> content::ServiceWorkerVersion::StartWorker()
-> (IPC from the browser to a renderer)
-> WebEmbeddedWorkerImpl::StartWorkerContext()
-> ServiceWorkerGlobalScope::FetchAndRunClassicScript()

(Note that renderer processes of above don't have to be the same.)

This CL has behavioral changes. The 'Referer' header of the request
triggered by register() in a document will be changed.
Before this CL it was the url of the top-level script but after this CL
it is document's url. This behavior is consistent with the spec.

Bug: 937177
Change-Id: I4c137d10b7d5204151ae7eff8bfd8fdeb34fb365