Support for Canvas and Text system color keywords
Add support for the new system color names, Canvas and Text. Canvas
maps to the Window system color, and Text maps to WindowText.

Text was already added as a system color previously, and was
mapped to a separate value than WindowText for Mac. This mapping
was left as-is for Mac.

Spec: https://drafts.csswg.org/css-color-4/#css-system-colors

Intent-to-ship: https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/-apVTp8jBi0

Bug: 1015968
Change-Id: I0047e6b6d6c87f13f556d5ad147d9c4be421a592